v-2-2-4
Fix to further isolate delayed actions with a new internal ID for added config commands so we can pinpoint exactly which ones get cancelled or cleaned up. Also, improve clean-up by not trying to run actions on elements that are not there - this was only partially working and not having this would have been affecting performance where a lot of cancel-timer actions were taking place.