Check-in prior to 2.4.0 release. Not recommended to use in production.
Fix to @command and @conditional so they can work in components (but remain common to all components and the document scope) and fix anomaly of them failing when declared later in the config after the command/conditional is actually run due to them being place in the body:init event in an incorrect sequence. Plus the create-command tests.