Check-in prior to 2.4.0 release.
Tests for scroll-into-view and var commands. Plus fix to convert tabs to spaces when parsing config, but only for non-HTML config. This is so that developers can multi-line their config and it won't just collapse with no spacing. Also fix to not remove double-quotes from string variables when they are passed as parameters into functions. _ACSSRepQuo corrected to remove only the outer double-quotes as was originally intended, as variables when evaluated lose escaping with quotes.