mainly SPA upgrade, but a few important fixes (see description)
Fix for a clean transition when adding and removing inline config.
Pass original event around when triggering.
SPA @pages upgrade to handle hash routes.
Offline SPA upgrade to handle template-based actual single-page app.
New url-replace command.
Correctly handle SPA nav attribute propagation when target is a child of the a tag.
Fix to scenario when the top-level render item was an inner table element, the table tags were disappearing.
Fix to ajax-pre-get sometimes getting files twice due to race condition.
Fix to observed attribute not working as a reactive variable if the attribute did not exist at the moment the element was rendered.