html-entity-decode option added to set-attribute and set-property commands + plus major escaping fix
This allows the handling of variables that have already been escaped on the server to not get effectively double escaped when they get added to an attribute and later rendered.

Plus *major* fix to do with escaping variables only at the point of render command and never earlier than that. This improves predictability in that variables should remain at a constant value despite re-assigning, until the point when actually rendered, which is where we actually only need the escaping for placing into the DOM to take place. Note that variable updates place directly into text nodes, so no specific escaping is needed there.